add some functionality to Vector, delete copy in AccumulatorSuite
